fix: adjust type casting for safe_limit and offset in generate_query function

+++ src/graphql/query/mod.rs
@@ -79,8 +79,8 @@ pub fn generate_query(table: Arc<Table>, pool: Arc<Pool>) -> Field {
                 }
 
                 let order_pairs = sql::parse_order_by(&order_by, &columns)?;
-                let safe_limit = first.unwrap_or(100).clamp(1, 1000) as i32;
-                let off = offset.unwrap_or(0).max(0) as i32;
+                let safe_limit = first.unwrap_or(100).clamp(1, 1000);
+                let off = offset.unwrap_or(0).max(0);
 
                 let (total_count, json_rows) = if !order_pairs.is_empty() {
                     let first_pair = &order_pairs[0];
